BACKPORT: mm: hardwire page_cluster = 1 for zram
Force page_cluster to 1 when ZRAM or ZSWAP is enabled instead of the
default value of 3. This reduces the number of pages read during swap-in
from 8 (2^3) to 2 (2^1), avoiding unnecessary I/O on compressed RAM
swap where sequential readahead provides no benefit.

A separate user_page_cluster variable is used for the sysctl interface
to preserve user-visible behavior while the actual swap clustering uses
the hardwired value.

Change-Id: I67624ff774550080ac5e3d58c046349e39f95530
Signed-off-by: psndna88 <psndna88@gmail.com>
Signed-off-by: Quince <quinceroms@gmail.com>