BACKPORT: binder: frozen notification
Frozen processes present a significant challenge in binder transactions. When
a process is frozen, it cannot, by design, accept and/or respond to binder
transactions. As a result, the sender needs to adjust its behavior, such as
postponing transactions until the peer process unfreezes. However, there is
currently no way to subscribe to these state change events, making it
impossible to implement frozen-aware behaviors efficiently.

Introduce a binder API for subscribing to frozen state change events. This
allows programs to react to changes in peer process state, mitigating issues
related to binder transactions sent to frozen processes.

Implementation details:
For a given binder_ref, the state of frozen notification can be one of the
followings:

1. Userspace doesn't want a notification. binder_ref->freeze is null.
2. Userspace wants a notification but none is in flight. list_empty(&binder_ref->freeze->work.entry) = true
3. A notification is in flight and waiting to be read by userspace. binder_ref_freeze.sent is false.
4. A notification was read by userspace and kernel is waiting for an ack. binder_ref_freeze.sent is true.

When a notification is in flight, new state change events are coalesced
into the existing binder_ref_freeze struct. If userspace hasn't picked up
the notification yet, the driver simply rewrites the state. Otherwise, the
notification is flagged as requiring a resend, which will be performed once
userspace acks the original notification that's inflight.

Signed-off-by: Yu-Ting Tseng <yutingtseng@google.com>
Signed-off-by: Quince <quinceroms@gmail.com>