BACKPORT: binder: allow freeze notification for dead nodes
Alice points out that binder_request_freeze_notification() should not return
EINVAL when the relevant node is dead. The node can die at any point even if
the user input is valid. Instead, allow the request to be allocated but skip
the initial notification for dead nodes. This avoids propagating unnecessary
errors back to userspace.

Signed-off-by: Carlos Llamas <cmllamas@google.com>
Signed-off-by: Quince <quinceroms@gmail.com>